requirements.txt
python-dotenv==1.0.0
streamlit
streamlit-extras
altair<5
openai
tiktoken
sentence-transformers==2.2.2
pdfminer.six
docx2txt
langchain 
langchainhub
streamlit-ext
chromadb==0.4.15
pandas
langchain-community
langchain-openai
faiss-cpu
psycopg2-binary
openpyxl
langchain_community